spark_train.py for Titanic
Some code I wrote about a year ago, which demonstrates the Titanic example from Kaggle using Spark and DecisionTrees. I have not tried to do any feature engineering, etc, over the dataset in any significant way, the purpose of this is just to show how to use Spark to do a classification problem.